# Improved performance of `bitmap::from_trusted` (3x)