Optimizes the creation and extending of `MutableBitmap`. It mostly re-arranges the code to leverage bitmap operations and the trustedLen invariant inside the hot loop.

```
Gnuplot not found, using plotters backend
bitmap from_trusted_len 2^10                                                                            
                        time:   [343.17 ns 344.70 ns 346.28 ns]
                        change: [-61.785% -61.465% -61.171%] (p = 0.00 < 0.05)

bitmap from_trusted_len 2^12                                                                             
                        time:   [1.2747 us 1.2815 us 1.2886 us]
                        change: [-63.765% -63.286% -62.869%] (p = 0.00 < 0.05)

bitmap from_trusted_len 2^14                                                                             
                        time:   [5.0600 us 5.0821 us 5.1083 us]
                        change: [-62.807% -62.486% -62.179%] (p = 0.00 < 0.05)

bitmap from_trusted_len 2^16                                                                             
                        time:   [20.037 us 20.147 us 20.256 us]
                        change: [-63.013% -62.660% -62.334%] (p = 0.00 < 0.05)

bitmap from_trusted_len 2^18                                                                            
                        time:   [80.056 us 80.641 us 81.273 us]
                        change: [-62.689% -62.358% -62.044%] (p = 0.00 < 0.05)

bitmap from_trusted_len 2^20                                                                            
                        time:   [317.06 us 318.74 us 320.50 us]
                        change: [-62.787% -62.497% -62.187%] (p = 0.00 < 0.05)

bitmap extend_from_trusted_len_iter 2^10                                                                            
                        time:   [454.37 ns 456.53 ns 458.83 ns]
                        change: [-54.421% -54.039% -53.670%] (p = 0.00 < 0.05)

bitmap extend_from_trusted_len_iter 2^12                                                                             
                        time:   [1.3878 us 1.3938 us 1.4003 us]
                        change: [-60.631% -60.227% -59.785%] (p = 0.00 < 0.05)

bitmap extend_from_trusted_len_iter 2^14                                                                             
                        time:   [5.1372 us 5.1573 us 5.1786 us]
                        change: [-61.752% -61.251% -60.609%] (p = 0.00 < 0.05)

bitmap extend_from_trusted_len_iter 2^16                                                                             
                        time:   [20.073 us 20.211 us 20.355 us]
                        change: [-62.762% -62.438% -62.100%] (p = 0.00 < 0.05)

bitmap extend_from_trusted_len_iter 2^18                                                                            
                        time:   [79.685 us 80.196 us 80.758 us]
                        change: [-62.818% -62.511% -62.186%] (p = 0.00 < 0.05)

bitmap extend_from_trusted_len_iter 2^20                                                                            
                        time:   [319.68 us 321.64 us 323.73 us]
                        change: [-62.872% -62.617% -62.346%] (p = 0.00 < 0.05)
```