@jorgecarleitao 2021-11-05
Optimized it a bit more (to 3x in total) by using chunks of 64 bits instead of chunks of 8. A tinny bit more complex, but 33% are worth imo. 🚀